third_party/blink/web_tests/virtual/text-antialias/find-soft-hyphen-expected.txt

Tests find for strings with soft hyphens in them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canFind("ab", "a" + softHyphen + "b") is true
PASS canFind("ab", "a" + softHyphen + softHyphen + "b") is true
PASS canFind("àb", "a" + softHyphen + "b") is true
PASS canFind("ab", "a" + softHyphen + "̀b") is true
PASS canFind("ab", "à" + softHyphen + "b") is true
PASS canFind("a" + softHyphen + "b", "a" + softHyphen + "b") is true
PASS canFind("a" + softHyphen + "b", "ab") is true
PASS canFind("a" + hyphen + "b", "a" + softHyphen + "b") is false
PASS canFind("a-b", "a" + softHyphen + "b") is false
PASS successfullyParsed is true

TEST COMPLETE
